+++ lib/models/place.dart
@@ -9,8 +9,10 @@ class Place{
   double? longitude;
   String? province;
   String? suburb;
+  String? streetNumber;
+  String? streetAddress;
 
-  Place({this.address,this.code, this.country,this.city,this.province,this.suburb, this.latitude, this.longitude});
+  Place({this.address,this.code, this.country,this.city,this.province,this.suburb, this.latitude, this.longitude, this.streetAddress, this.streetNumber});
 
 
   Place.newInstance(this.address, this.longitude, this.latitude);
@@ -24,6 +26,8 @@ class Place{
     suburb = json['suburb'];
     latitude = json['latitude'];
     longitude = json['longitude'];
+    streetNumber = json['streetNumber'];
+    streetAddress = json['streetAddress'];
 
   }
   Map<String, dynamic> toJson() {
@@ -51,6 +55,12 @@ class Place{
     }
     if (suburb != null) {
       data['suburb'] = suburb;
+    }
+    if (streetNumber != null) {
+      data['streetNumber'] = streetNumber;
+    }
+    if (streetAddress != null) {
+      data['streetAddress'] = streetAddress;
     }
     return data;
   }